+++ lib/td_dd_web/controllers/grant_request_filter_controller.ex
@@ -8,21 +8,9 @@ defmodule TdDdWeb.GrantRequestFilterController do
   def search(conn, params) do
     claims = conn.assigns[:current_resource]
 
-    fixed_params = maybe_fix_approved_params(params)
+    params = Search.apply_approve_filters(params)
 
-    {:ok, filters} = Search.get_filter_values(claims, fixed_params)
+    {:ok, filters} = Search.get_filter_values(claims, params)
     render(conn, "show.json", filters: filters)
   end
-
-  defp maybe_fix_approved_params(
-         %{"filters" => %{"must_not_approved_by" => approved_by} = filters} = params
-       ) do
-    must_without_approved = Map.delete(filters, "must_not_approved_by")
-
-    params
-    |> Map.put("filters", must_without_approved)
-    |> Map.put("must_not", %{"approved_by" => approved_by})
-  end
-
-  defp maybe_fix_approved_params(params), do: params
 end