+++ lib/td_dq/rules/rule_results/policy.ex
@@ -6,27 +6,27 @@ defmodule TdDq.Rules.RuleResults.Policy do
 
   @behaviour Bodyguard.Policy
 
-  def authorize(:query, %{role: "user"} = claims, _params),
+  def authorize(_action, %{role: "admin"}, _params), do: true
+  def authorize(_action, %{role: "service"}, _params), do: true
+
+  def authorize(:query, %{} = claims, _params),
     do: Permissions.authorized?(claims, :view_quality_rule)
 
-  def authorize(:view, %{role: "user"} = claims, _params),
+  def authorize(:view, %{} = claims, _params),
     do: Permissions.authorized?(claims, :view_quality_rule)
 
-  def authorize(:upload, %{role: "user"} = claims, _params),
+  def authorize(:upload, %{} = claims, _params),
     do: Permissions.authorized?(claims, :manage_rule_results)
 
   def authorize(
         :manage_remediations,
-        %{role: "user"} = claims,
+        %{} = claims,
         %RuleResult{
           implementation: %{domain_id: domain_id}
         }
       ) do
     Permissions.authorized?(claims, :manage_remediations, domain_id)
   end
-
-  def authorize(_action, %{role: "admin"}, _params), do: true
-  def authorize(_action, %{role: "service"}, _params), do: true
 
   def authorize(_action, _claims, _params), do: false
 end